[refactor] #9  repository 폴더 추가
