+++ src/main/java/org/netway/dongnehankki/user/application/UserService.java
@@ -28,7 +28,7 @@ import org.springframework.security.config.annotation.authentication.builders.Au
 import org.springframework.security.core.Authentication;
 
 import org.netway.dongnehankki.store.domain.Store;
-import org.netway.dongnehankki.store.infrastructure.StoreRepository;
+import org.netway.dongnehankki.store.infrastructure.repository.StoreRepository;
 
 @Service
 @RequiredArgsConstructor
